requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
clearbit==0.1.7
cssselect==1.1.0
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
fake-useragent==0.1.11
idna==2.10
jsonfield==3.1.0
lxml==4.6.2
parse==1.19.0
pyee==8.1.0
pyhunter==1.7
PyJWT==2.0.1
pyppeteer==0.2.5
pyquery==1.4.3
pytz==2021.1
requests==2.25.1
requests-html==0.10.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tqdm==4.57.0
urllib3==1.26.3
w3lib==1.22.0
websockets==8.1
